refactor: 폴더 구조 개선 및 올바르지 않는 컴포넌트 파일명 수정
