+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next"
 import { Inter } from "next/font/google"
 import "./globals.css"
 import Header from "../components/Header"
-import WebSocketProvider from "@/components/WebSocketProvider"
+import WebSocketProvider from "@/providers/WebSocketProvider"
 
 const inter = Inter({ subsets: ["latin"] })
