+++ src/components/CumulativeAskingPrice.tsx
