+++ src/components/GeneralAskingPrice.tsx
