+++ src/components/GeneralAskingTotalPrice.tsx
